# i18n: Avoid rebuilding the Translations map for every lookup